Fix label color bug in MainGameView plus small improvements

+++ app/src/main/java/wazzle/controller/maingame/MainGameControllerImpl.java
@@ -141,7 +141,7 @@ public final class MainGameControllerImpl implements MainGameController {
 	 */
 	@Override
 	public String highestScoreWord() {		
-        double max = this.game.get()
+        int max = this.game.get()
                 .wordsFound()
                 .stream()
                 .map(w -> this.game.get().getScoreFromWord(w))
